devops - structure out daily and weekly scripts
