+++ .github/workflows/weekly-product.yml
@@ -31,7 +31,7 @@ jobs:
       - run: npm i
 
       - name: Post weekly digest
-        run: node index.js product
+        run: node weekly.js product
         env:
           GITHUB_TOKEN: ${{ secrets.CD_GITHUB_TOKEN }}
           SLACK_TOKEN: ${{ secrets.SLACK_TOKEN }}
