+++ package.json
@@ -3,8 +3,16 @@
   "version": "0.1.0",
   "description": "Retrieve updates based on activities happening across the GitHub organization",
   "main": "index.js",
+  "source": "index.js",
+  "files": [
+    "index.js",
+    "sources",
+    "targets",
+    "README.md",
+    "LICENSE"
+  ],
   "scripts": {
-    "weekly": "env $(print-env) node index.js",
+    "weekly": "env $(print-env) node weekly.js",
     "daily": "env $(print-env) node daily.js",
     "lint": "eslint .",
     "test": "echo \"Error: no test specified\" && exit 1"
