+++ weekly.js
@@ -1,7 +1,7 @@
 const { DateTime } = require('luxon')
 
 const {
-  sources: { github: { issuesByRange, enrichIssues, ignoreProjects, ignoreBotUsers, formatPreviously } },
+  sources: { github: { issuesByRange, enrichIssues, ignoreProjects, ignoreBotUsers, formatDigest } },
   targets: { slack: { uploadMD } },
 } = require('.')
 
@@ -9,24 +9,24 @@ const { ORG_TZ = 'America/Toronto' } = process.env
 const stripMS = (dt) => `${dt.toISO().split('.')[0]}Z`
 
 
-const dailyPreviously = () => {
-  // last work day in ISO string - ms portion
-  const today = DateTime.utc().startOf('day').setZone(ORG_TZ, { keepLocalTime: true })
-  const lastYst = today.minus({ days: today.weekday === 1 ? 3 : 1 })
-  const yst = today.minus({ day: 1 })
+const weeklyDigest = () => {
+  const team = process.argv[2]
+  // weekly range in ISO string but drops ms portion
+  const yst = DateTime.utc().minus({ day: 1 }).setZone(ORG_TZ, { keepLocalTime: true })
+  const lastYst = yst.minus({ week: 1 }).plus({ day: 1 })
   const start = stripMS(lastYst.startOf('day').toUTC())
   const end = stripMS(yst.endOf('day').toUTC())
 
   issuesByRange({ start, end })
     .then((issues) => issues.filter(ignoreProjects))
     .then((issues) => issues.filter(ignoreBotUsers))
-    .then((issues) => enrichIssues({ issues, start, end }))
-    .then(formatPreviously)
+    .then((issues) => enrichIssues({ issues, start, end, team }))
+    .then(formatDigest)
     .then(uploadMD())
     .then(console.log)
     .then(console.error)
 }
 
 if (require.main === module) {
-  dailyPreviously()
+  weeklyDigest()
 }
